move /auth/google and / route to routes directory
