+++ backend/index.ts
@@ -7,6 +7,6 @@ const server = http.createServer(app)
 
 const port = 8000;
 
-app.listen(8000, () => {
+server.listen(port, () => {
     console.log(`Server started: http://localhost:${port}/`);
 });
